refactor adaptor
